+++ test/add-info.spec.js
